fix(1804): make the close outcome identity-stable and never downgrade a recovery
Two real defects the full-suite run surfaced (both green in isolation, red at
5,100 tests — the pollution class the dossier warned about).

1. ActivityCloseOutcome moves db/activities.py -> models.py. Callers compare it
   by `is`, and db_harness evicts `db.activities` from sys.modules per test. Any
   test that imported services.activity_service BEFORE an eviction held one enum
   class while a later `from db.activities import ...` got a different one, so
   every identity check silently went False. models is the leaf everything
   imports and nothing re-imports, so the object stays the same one — and it now
   sits beside ActivityState/TaskExecutionStatus, where a shared db↔service
   contract type belongs. Re-exported through db.activities.

2. Both _recover_execution closes get their own try/except. The terminal write
   and the capacity release have already succeeded at that point; letting a
   close failure propagate flipped a recovered row into the `errors` bucket and
   would make the watchdog retry a row it had already recovered. Fail-open is
   the rule for everything after a committed terminal — this was the one place I
   left it implicit.

Full unit suite: 5136 passed, 14 skipped, 0 failed.
